Adds ability to specify user to login as.
The goal of this commit is to make it easier for people to get started with Keywhiz. It enables us to
provide a set of instructions in the docs that can be copy-pasted as-is. Right now, this instructions
would be:
mvn package -am -pl cli
java -Djavax.net.ssl.trustStore=server/src/main/resources/dev_and_test_truststore.p12
     -Djavax.net.ssl.trustStorePassword=ponies -Djavax.net.ssl.trustStoreType=PKCS12
     -jar ./cli/target/keywhiz-cli-*-SNAPSHOT-shaded.jar --user keywhizAdmin <cmd>

The session now is saved in ~/.keywhiz.<username>.cookie. This ensures that
if you login as user1 and then call the cli with user2, you won't be automatically
logged-in as user1.