Wrap OffsetDateTime in ApiDate & update Jackson.
Changes:

Update Jackson to the newest stable release (2.6.1).

We wrap OffsetDateTime in an ApiDate that provides a thin wrapper around it
and custom Json de/serializers.  It also has static now() and parse() methods
that are heavily used in tests, and equality & hashcode checks implented in the
obvious way.

There's a lot of mundane changes to pipe the type throughout the API, and to
update tests appropriately.

Remove the now unused options to set up the date format in Jackson: This makes
it easier to find buggy serialization, as they get written as a number instead.

- Background & Alternatives:

Keywhiz uses Jackson for JSON de/serialization.  When OffsetDateTime is
serialized with Jackson's JSR310 module, which omits the milliseconds if
they're all zero.  While this is unlikely to happen, as most times in the
keywhiz world are timestamps and the milliseconds are effectively random, it
does make Keywhiz-fs unhappy, and unable to list secrets, or read the secret
with the unsupported date format.

We set the jackson date format explicitly, but that's a red herring: The JSR310
date module doesn't use this.  We could explicitly set the pattern with
@JsonPattern to be yyyy-MM-dd'T'mm-HH-ss.SSS'Z' which serializes dates properly,
but then won't deserialize because OffsetDateTime's constructor doesn't know
what timezone it's in.  If we used the unquoted 'Z', the timezone will serialize
as +0000, which is undesired too, but it deserializes properly.

We could load a custom module into Jackson that provides custom de/serializers
for OffsetDateTime, but we need to load the Jdk8 module, which provides
conflicting de/serializers for OffsetDateTime.  You could probably get that to
work with the right ordering, but that feels awfully fragile to me (and I
didn't get it to work).