Update automation endpoints to stop reusing deleted secrets' database entries.
Keywhiz initially reused database rows when a secret was deleted and
a second secret with the same name was created; this meant that the
second secret had the same name and ID as the first secret. This made
it impossible, when listing the versions for a secret, to distinguish
versions made before a secret's deletion from versions created after
the deletion.  Once the ability to roll back to a previous version was
added, this ambiguity could potentially leak secret contents from the
first secret's owners to the second secret's owners.

Note that this commit does not update the admin client, which still lists
deleted secrets among its versions.