Update SanitizedSecret to use content for update time
When SanitizedSecret is used to represent a secret version, it
is constructed from a secret series and a secret content record.
Originally, the SanitizedSecret's creation and update times were
taken from the secret series; however, this loses information on
when the version was created.  This commit keeps the SanitizedSecret's
creation date set to the series' creation date, but changes its
update time to match the current content record's update time.  If
the content is the secret's current version, this does not change
SanitizedSecret's behavior; however, if the content is a past version
the SanitizedSecret's update time reflects when that version was
last updated/created (since creation and update are always identical
for a secret content record).