Synchronize timestamps for secret creation and updates.
Secrets are represented in two tables, one for the "secret series"
and one for each set of contents that the secret has.  Both the
secret and its contents have update and creation timestamps.
This commit ensures that these timestamps will always match
(i. e. the first content for a secret will have a createdAt value
matching the createdAt value in the secret table, and immediately
after a secret update which creates new content the secret's
updatedAt value will match the updatedAt/createdAt value for its
current content, which was just created).

This should fix an intermittent failure in the AclDAO test, which
compares a SanitizedSecret whose timestamps use the values from the
"secrets" table against a SanitizedSecret (for the same secret) whose
timestamps use the values from the "secrets_content" table for the
secret's current contents.  Occasionally, the timestamp from the
"secrets" table (which is updated first) was a second behind the
timestamp from the "secrets_content" table.