Add fields to track content creation information in Secret and SanitizedSecret
If secrets have had their contents reset to an earlier version,
or when SanitizedSecret is used to represent an older version of a
secret, the content creation information will differ from the
secret's creation and update data.  This adds a field to explicitly
track content creation data (which is equivalent to content update
data since Keywhiz never updates content records).