Downgrade logback version
Dropwizard before version 1.0.0 is incompatible with logback
after 1.1.3, because logback 1.1.5 and 1.1.6 added checks to
RollingFileAppender which cause errors in Dropwizard's
FileAppenderFactory (see Dropwizard issue 1478 on Github).  The
result is that Dropwizard before 1.0.0 will log to the console but
not to files if logback is after 1.1.3.  Keywhiz currently uses
Dropwizard 0.8.0, so logback must be kept to an early version
until it can be upgraded together with Dropwizard.