Add command to remove database rows for deleted secrets
Deleted secrets in Keywhiz are still present in the database;
the only difference between a deleted secret and an existing secret
is that `secrets.current` for a deleted secret is null, while it's
the ID of a row in the `secrets_content` table for an existing
secret.

This adds a command to remove database rows for secrets which have
been deleted before an input date.  This command must be treated
extremely carefully, since its effects are permanent unless the
database is restored from a backup.

This command will be valuable in removing unused deleted secrets
since a recent commit changed the naming pattern for deleted secrets.