Make lower bound on retrieving expiring secrets exclusive
This matches the existing behavior of this endpoint (exclusive on
a lower bound of "now" for the expiration dates of secrets to
retrieve).  More importantly, it makes constructing a sliding
window of expirations to retrieve significantly easier for clients,
since consecutive sections of the window no longer overlap (i. e.
[t1 -> t2] and [t2 -> t3] will not overlap).