Move Keywhiz client creation through admin API to v2 request
The "automation.v2" creation requests are more detailed and widely
used than the original creation request (which is now used only
in the deprecated automation API and the admin API). This moves
client creation to use the v2 creation request format, so that
future updates to client creation only need to change one class.

The client response object is not migrated to the "automation.v2"
format, since the response object--which combines a client with
its groups and secrets--is specialized for the admin API.