Move Keywhiz group creation through admin API to v2 request
The "automation.v2" creation requests are identical in their
JSON representation to the original creation request; the only
difference is that the "automation.v2" request uses a builder
pattern to create the Java class. This changes group creation
to use the v2 creation request format, so that future updates
to group creation only need to change one class.

The group response object is not migrated to the "automation.v2"
format, since the response object is specialized for the admin API.