Add support for authenticating clients through Envoy-format XFCC header
Envoy, a widely used L7 proxy, sets the x-forwarded-client-cert
header to contain the certificate for clients connecting to
the proxy. It then can connect directly to its server (such as
Keywhiz) using TLS.

This updates Keywhiz to parse client certificates out of the
XFCC header if they are provided. If the XFCC header was provided,
Keywhiz ignores any certificates from the direct connection; this
is necessary to avoid using the proxy (if it is also a Keywhiz
client) as the client when taking action rather than the real
client which connected to the proxy.

This still has some limitations, which may be addressed in future
commits:
* the XFCC header must be formatted as defined by Envoy
8 the XFCC header must have only one certificate in it, which corresponds
to the certificate for the real Keywhiz client
* only the "Cert" field from the XFCC header is parsed for client
data, although other fields (such as "Subject" and "URI") could carry
client data without the need for parsing a certificate
* traffic with the XFCC header set is blocked based on a port, but
not based on client identity; Keywhiz could instead require traffic
to come from a fixed port and from an allowed, verified proxy client,
to further reduce the risk of a client setting the XFCC
header to steal another client's secrets.

This commit also introduces a configuration option to control
whether SPIFFE URIs are used to authenticate clients, but the
option is not yet in use.