Only allow x-forwarded-client-cert header from certain clients
Since setting the x-forwarded-client-cert header could allow a
client to impersonate another Keywhiz client, only allow a small
set of clients to use this header. This should be reserved for
clients that are known to set the header while acting as proxies
for other clients.

A later PR will extend this to support identifying clients by
SPIFFE IDs as well as client name (both when allowing the xfcc header
and generally when identifying clients).