Allow traffic to set x-forwarded-client-cert header based on clients only
Allowing all traffic from a port to set the x-forwarded-client-cert
header for identifying Keywhiz tickets introduces significant
security challenges around ensuring that only a restricted group
of clients (such as a proxy) can access this port. Restricting the
entities that can set the XFCC header to only a specific set
of clients is more secure by default.

This also adjusts the ClientAuthFactory test to better distinguish
the client that connected to Keywhiz from the client for which
Keywhiz should take action (from the XFCC header).