Refactor client authorization for clarity and future SPIFFE work
This refactors the client authorization flow to clarify the
distinction between the security context of a request (which may
have been made directly by a client or by a proxy acting on behalf
of a client) and the actual client request information (parsed from
the request security context or from the x-forwarded-client-cert
header).

This includes an outline of supporting SPIFFE as an identifier for
both the requestor (client or proxy) and the client. Actually
adding this support will be handled in a later commit.