Designate ports to receive all or no traffic with XFCC header set
In addition to allowing only certain clients to set the
x-forwarded-client-cert header, this configures Keywhiz to either
require all traffic from a port to provide client identity through
the x-forwarded-client-cert header, or else require no traffic to
that port to set the x-forwarded-client-cert header. Without this
check, a bug in a proxy that caused the x-forwarded-client-cert header
to be invalid would cause Keywhiz to fall back to using the proxy's
certificates as the identifiers for the Keywhiz client. After this
change, traffic from the proxy (on the proxy's designated port)
without an XFCC header set will be rejected.