Retrieve clients by SPIFFE ID and refactor client authentication
This retrieves clients based on the SPIFFE ID parsed from
the client information.

This also refactors the client authentication factory to separate
extracting information about the requestor from a request (either
the mTLS data associated with the request, or the
x-forwarded-client-cert header on a request), and retrieving a
Keywhiz client based on that extracted information.