Use HttpServletRequest to check XFCC port
The ContainerRequest constructed by Jersey does not include information
about where the request was received by Keywhiz; the port in its URI
is derived from the port (if any) specified by the original requestor.
The HttpServletRequest, however, identifies which Keywhiz port traffic
came to, which allows us to enforce controls--notably whether certificates
are allowed to be set by the x-forwarded-client-cert header--based on
the port associated with incoming traffic.