Improve handling of empty secret contents on update
When empty secret contents are passed to the keywhiz.cli
command-line tool, it silently does not update the secret's
contents. This updates the command-line tool to error when
empty contents are passed but the --content flag is specified.

This also updates Keywhiz' internal update endpoint to check
for empty contents, since that endpoint is not accessed
through the SecretController's Builder class (unlike the create
and complete-update endpoints, which use SecretController's
validation to ensure that secret contents are present).