Add index on creation time for secret contents
Multiple endpoints that work with secrets' versions, notably
the endpoint to list all secret versions and the check for
old versions of a secret when creating or updating a secret,
order secret versions by their creation date.

Without this index, a secret with thousands of versions takes
seconds to update because the attempt to prune old versions of
the secret before updating takes a second or more to list and
order the existing versions.