Force expiry index use when retrieving expiring secrets
When the query to retrieve expiring secrets examines a large
number of rows (the cutoff in one series of tests appeared to
be approximately a million rows), it may fail to use the index
on secrets_content.expiry. This results in a near-full table scan
and a dramatic increase in the time required to run this query.

This change forces the optimizer to use the expiration index
when the query will scan a range of expiration dates. It does not
affect queries that do not use the secrets' expiration in the
"where" clause.