Fix tests which rely on ordering of JSON fields
Since the ordering of fields in a serialized JSON object
is not deterministic, tests which serialize an object
and compare it to a fixed file will be flaky. This replaces
such tests with 1) a test that a fixed file deserializes
to the expected object and 2) a test that a round-trip
serialization and deserialization results in the original
object.